Add navigation old `view.js` files back
