+++ packages/block-library/package.json
@@ -68,6 +68,7 @@
 		"fast-average-color": "^9.1.1",
 		"fast-deep-equal": "^3.1.3",
 		"memize": "^2.1.0",
+		"micromodal": "^0.4.10",
 		"remove-accents": "^0.4.2",
 		"uuid": "^8.3.0"
 	},
